+++ tests/newspack-ai-newsletter-test-config.php
@@ -13,9 +13,11 @@
 return [
 	'base_directory'    => '/tmp/newspack-ai-newsletter',
 	'num_partitions'    => 1,
-	'num_segments'      => 2,
 	'segment_size'      => 1024,
-	'max_lifespan'      => 0,
+	'min_segments'      => 2,
+	'max_segments'      => 2,
+	'min_lifetime'      => 0,
+	'max_lifetime'      => 0,
 	'memcache_servers'  => [],
 	'enable_logging'    => false,
 	'enable_jobs'       => false,